// ExistingCode/src/data-access/domain/tables/assaybatchdetail/schema.ts
export const AssayBatchDetailTableInsertBaseSchema = z.object({
	/**
	 * dataType: uniqueidentifier
	 * defaultValue: (newid())
	 */
	assayBatchDetailId: z.any().optional(),
	/**
	 * dataType: nvarchar
	 */
	labCode: z.string().max(20),
	/**
	 * dataType: nvarchar
	 */
	batchNo: z.string().max(50),
	/**
	 * dataType: nvarchar
	 */
	originalMethod: z.string().max(50),
	/**
	 * dataType: nvarchar
	 */
	labElement: z.string().max(50),
	/**
	 * dataType: nvarchar
	 */
	element: z.string().max(50).nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	repeat: z.string().max(10).nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	unitCode: z.string().max(50).nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	genericMethod: z.string().max(50).nullable().optional(),
	/**
	 * dataType: datetimeoffset
	 * defaultValue: (sysdatetimeoffset())
	 */
	createdOnDt: z.any().optional(),
	/**
	 * dataType: nvarchar
	 * defaultValue: (suser_sname())
	 */
	createdBy: z.string().max(50).optional(),
	/**
	 * dataType: datetimeoffset
	 */
	modifiedOnDt: z.any().nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	modifiedBy: z.string().max(50).nullable().optional(),
	/**
	 * dataType: timestamp
	 */
	rv: z.date(),
});